fix readme and install script
